ci(sandbox): GHCR-only image names + main→beta / tag→stable pipelines
Bundles the in-flight security-audit work plus this session's CI/CD
fixes onto fix/security-audit-v0.0.8.

## Why publishing failed before

The runtime aliased pulled images (`ghcr.io/.../temps-sandbox-node:0.1.0-beta`)
to a local short name (`temps-sandbox-node:0.1.0`), so `docker ps` showed
`:0.1.0` even though no `:0.1.0` ref ever existed on GHCR. Combined with
the release workflow only firing on tags that hadn't been cut for sandbox
content, hosts ended up on stale local builds.

## Image name refactor (crates/temps-agents/src/sandbox/docker.rs)

`image_name_for_runtime` now returns the full GHCR name end-to-end
including channel suffix. No more rename step. Added
`runtime_from_image_name` for recovery. Removed `hub_image_for_runtime`
and `try_pull_and_tag`.

## Dockerfile generator fixes

- `bun add -g @openai/codex` requires Node (codex is a Node script).
  Every base now installs `nodejs` in extra_packages.
- `oven/bun:latest` bakes `BUN_INSTALL_BIN=/usr/local/bin` into the
  image env, making `bun add -g` fail with EACCES under the
  unprivileged `temps` user. Pin both `BUN_INSTALL` and
  `BUN_INSTALL_BIN`.

## Pipelines

- `release.yml` (on `v*.*.*`): STABLE — `:<version>`, `:<version>-stable`,
  `:latest`, `:stable`, `:<sha>`. Now also covers preview-gateway with
  the channel split (was always `:latest`).
- `sandbox-images-beta.yml` (new, on main push): BETA — `:<version>-beta`,
  `:beta`, `:<sha>`. Path filter prevents server-only changes from
  triggering 20 min image rebuilds.

Both support `workflow_dispatch` with `dry_run`. Manual dispatches
always treat themselves as prerelease.

## Build script + helper

`scripts/build-sandbox-images.sh` runs `print_bundle` to materialize
pty-agent + git-credential sources. `print_bundle` example now does
both bundles.

## Bundled in-flight work

Lands the in-progress security-audit work the workflow fix depends on:
new `temps-git-credential` crate (helper + daemon), sandbox path
centralisation, in-sandbox CLI auto-auth. Includes mock impl for the
new `GitProviderManagerTrait::mint_scoped_repo_token` method in test
code.

+++ crates/temps-entities/src/external_service_backups.rs
@@ -13,7 +13,7 @@ pub struct Model {
     pub state: String,
     pub started_at: DBDateTime,
     pub finished_at: Option<DBDateTime>,
-    pub size_bytes: Option<i32>,
+    pub size_bytes: Option<i64>,
     pub s3_location: String,
     pub error_message: Option<String>,
     pub metadata: Json,
